requirements.txt
pandas>=2.0
numpy>=1.24
scipy>=1.10
torch>=2.0
faiss-cpu>=1.7
httpx>=0.24
fastapi>=0.100
uvicorn>=0.22
sentence-transformers>=2.2
tqdm>=4.65
pyarrow>=12.0
aiosqlite>=0.19
